+++ pyArango/theExceptions.py
@@ -1,4 +1,4 @@
-class ArrangocityException(Exception) :
+class pyArangoException(Exception) :
 	"""The calss from witch all Exceptions inherit"""
 	def __init__(self, message, errors = {}) :
 		Exception.__init__(self, message)
@@ -7,44 +7,44 @@ class ArrangocityException(Exception) :
 	def __str__(self) :
 		return self.message + ". Errors: " + str(self.errors)
 
-class ConnectionError(ArrangocityException) :
+class ConnectionError(pyArangoException) :
 	"""Something went wrong with the connection"""
-	def __init__(self, message, URL, errors = {}) :
-		mes = "%s. URL: %s" % (message, URL)
-		ArrangocityException.__init__(self, mes, errors)
+	def __init__(self, message, URL, statusCode="", errors = {}) :
+		mes = "%s. URL: %s, status: %s" % (message, URL, statusCode)
+		pyArangoException.__init__(self, mes, errors)
 
-class CreationError(ArrangocityException) :
+class CreationError(pyArangoException) :
 	"""Something went wrong when creating something"""
 	def __init__(self, message, errors = {}) :
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class UpdateError(ArrangocityException) :
+class UpdateError(pyArangoException) :
 	"""Something went wrong when updating something"""
 	def __init__(self, message, errors = {}) :
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class DeletionError(ArrangocityException) :
+class DeletionError(pyArangoException) :
 	"""Something went wrong when deleting something"""
 	def __init__(self, message, errors = {}) :
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class TraversalError(ArrangocityException) :
+class TraversalError(pyArangoException) :
 	"""Something went wrong when doing a graph traversal"""
 	def __init__(self, message, errors = {}) :
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class ValidationError(ArrangocityException) :
+class ValidationError(pyArangoException) :
 	"""Something went wrong when validating something"""
 	def __init__(self, message, errors = {}) :
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class SchemaViolation(ArrangocityException) :
+class SchemaViolation(pyArangoException) :
 	"""Raised when someone tries to add a new field to an object belonging a to a Collection with enforced schema"""
 	def __init__(self, collection, field, errors = {}) :
 		message = "Collection '%s' does not have a field '%s' in it's schema" % (collection.__name__, field)
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class InvalidDocument(ArrangocityException) :
+class InvalidDocument(pyArangoException) :
 	"""Raised when a Document does not respect schema/validation defined in its collection"""
 	def __init__(self, errors) :
 		message = "Unsuccesful validation" 
@@ -53,38 +53,38 @@ class InvalidDocument(ArrangocityException) :
 			self.strErrors.append("%s -> %s" % (k, v))
 		self.strErrors = '\n\t'.join(self.strErrors)
 
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 	
 	def __str__(self) :
 		return self.message + ":\n\t" + self.strErrors
 
-class SimpleQueryError(ArrangocityException) :
+class SimpleQueryError(pyArangoException) :
 	"""Something went wrong with a simple query"""
 	def __init__(self, message, errors = {}) :
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class QueryError(ArrangocityException) :
+class QueryError(pyArangoException) :
 	"""Something went wrong with an aql query"""
 	def __init__(self, message, errors = {}) :
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class AQLQueryError(ArrangocityException) :
+class AQLQueryError(pyArangoException) :
 	"""Something went wrong with an aql query"""
 	def __init__(self, message, query, errors = {}) :
 		message = "Error in: %s.\n->%s" % (query, message)
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class CursorError(ArrangocityException) :
+class CursorError(pyArangoException) :
 	"""Something went wrong when trying to fetch data with a cursor"""
 	def __init__(self, message, cursorId, errors = {}) :
 		message = "Unable to retreive data for cursor %s: %s" % (cursorId, message)
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
-class TransactionError(ArrangocityException) :
+class TransactionError(pyArangoException) :
 	"""Something went wrong with a transaction"""
 	def __init__(self, message, action, errors = {}) :
                 message = "Error in: %s.\n->%s" % (action, message)
-		ArrangocityException.__init__(self, message, errors)
+		pyArangoException.__init__(self, message, errors)
 
 class AbstractInstanciationError(Exception) :
 	"""Raised when someone tries to instanciate an abstract class"""